+++ bigdata_briefs/api/utils.py
@@ -2,6 +2,10 @@ from typing import Type
 
 from pydantic import BaseModel
 
+from bigdata_briefs.api.examples import EXAMPLE_REPORT, EXAMPLE_STATUS
+from bigdata_briefs.api.sql_models import SQLWorkflowStatus
+from bigdata_briefs.sql_models import SQLBriefReport
+
 
 def get_example_values_from_schema(schema_model: Type[BaseModel]) -> dict:
     """
@@ -24,3 +28,12 @@ def get_example_values_from_schema(schema_model: Type[BaseModel]) -> dict:
         else:
             example_values[field_name] = field.default
     return example_values
+
+
+def status_report_example_models() -> tuple[SQLWorkflowStatus, SQLBriefReport]:
+    """
+    Returns a tuple with example instances of SQLWorkflowStatus and SQLBriefReport models.
+    To use for initialization of the database with example data.
+    """
+
+    return EXAMPLE_STATUS, EXAMPLE_REPORT