+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.0.0] - Unreleased
+## [3.1.0] - 2025-10-10
+
+### Added
+- Added example at db initialization
+- Improved FE
+
+### Changed
+- Modified topics list to be more user friendly
+- JS scripts moved to static/scripts/
+- Modified default watchlists
+
+## [3.0.0] - 2025-09-30
 
 ### Added
 - Added new models to separate `RetrievedSources` from `ReportedSources`, allowing to keep track of all retrieved sources while only including the used ones in the final report.